# frontend/src/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Planty API",
    "description": "A REST API for Planty - tracking plant care and growth metrics",
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/plants": {
      "get": {
        "tags": [
          "plants"
        ],
        "operationId": "list_plants",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of plants to return",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of plants to skip",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search term for plant names",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort order: date_asc, date_desc, name_asc, name_desc",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "includeArchived",
            "in": "query",
            "description": "Include archived plants in list results",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of plants",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "post": {
        "tags": [
          "plants"
        ],
        "operationId": "create_plant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Plant created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttributeSource": {
        "type": "string",
        "enum": [
          "identification",
          "coach",
          "user"
        ]
      },
...
      "CareTask": {
        "type": "object",
        "description": "A care task represents a specific type of care activity for a plant.\nExamples: \"Water\", \"Fertilize\", \"Prune leaves\", \"Repot\", \"Rotate\", \"Mist\", etc.\nEach can optionally have a recurring schedule (interval_days).",
        "required": [
          "id",
          "plantId",
          "userId",
          "name",
          "sortOrder",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "amount": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "archivedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "intervalDays": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "lastPerformed": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "plantId": {
            "type": "string",
            "format": "uuid"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          },
          "unit": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "CareTaskWithStatus": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CareTask"
          },
          {
            "type": "object",
            "required": [
              "isDue"
            ],
            "properties": {
              "daysOverdue": {
                "type": "integer",
                "format": "int64",
                "description": "How many days overdue (negative = days until due, 0 = due today, positive = overdue)",
                "nullable": true
              },
              "isDue": {
                "type": "boolean",
                "description": "Whether the task is currently due"
              },
              "nextDue": {
                "type": "string",
                "format": "date-time",
                "description": "When this task is next due (None if no interval or never performed with no interval)",
                "nullable": true
              }
            }
          }
        ],
        "description": "Computed view of a care task with due-date info for the frontend"
      },
...
      "CreateCustomMetricRequest": {
        "type": "object",
        "required": [
          "name",
          "unit",
          "dataType"
        ],
        "properties": {
          "dataType": {
            "$ref": "#/components/schemas/MetricDataType"
          },
          "name": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          }
        }
      },
...
      "CreatePlantAttributeRequest": {
        "type": "object",
        "required": [
          "key",
          "label",
          "value"
        ],
        "properties": {
          "category": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "key": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "source": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AttributeSource"
              }
            ],
            "nullable": true
          },
          "value": {
            "type": "string"
          }
        }
      },
      "CreatePlantCareTaskInput": {
        "type": "object",
        "description": "Inline care task definition used when creating a plant",
        "required": [
          "name"
        ],
        "properties": {
          "amount": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "intervalDays": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "lastPerformed": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "unit": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CreatePlantRequest": {
        "type": "object",
        "required": [
          "name",
          "genus"
        ],
        "properties": {
          "attributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreatePlantAttributeRequest"
            },
            "description": "Initial plant attributes (requirements/characteristics) to create with the plant",
            "nullable": true
          },
          "careTasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreatePlantCareTaskInput"
            },
            "description": "Initial care tasks to create with the plant",
            "nullable": true
          },
          "customMetrics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateCustomMetricRequest"
            },
            "nullable": true
          },
          "genus": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "CustomMetric": {
        "type": "object",
        "required": [
          "id",
          "plantId",
          "name",
          "unit",
          "dataType"
        ],
        "properties": {
          "dataType": {
            "$ref": "#/components/schemas/MetricDataType"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "plantId": {
            "type": "string",
            "format": "uuid"
          },
          "unit": {
            "type": "string"
          }
        }
      },
...
      "MetricDataType": {
        "type": "string",
        "enum": [
          "Number",
          "Text",
          "Boolean"
        ]
      },
...
      "PlantAttribute": {
        "type": "object",
        "required": [
          "id",
          "plantId",
          "key",
          "label",
          "value",
          "source",
          "sortOrder",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "category": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "key": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "plantId": {
            "type": "string",
            "format": "uuid"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          },
          "source": {
            "$ref": "#/components/schemas/AttributeSource"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "value": {
            "type": "string"
          }
        }
      },
...
      "PlantResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "genus",
          "customMetrics",
          "careTasks",
          "attributes",
          "createdAt",
          "updatedAt",
          "userId"
        ],
        "properties": {
          "archivedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "attributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlantAttribute"
            }
          },
          "careTasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CareTaskWithStatus"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "customMetrics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomMetric"
            }
          },
          "genus": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "previewId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "previewUrl": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "userId": {
            "type": "string"
          }
        }
      },
      "PlantsResponse": {
        "type": "object",
        "required": [
          "plants",
          "total",
          "limit",
          "offset"
        ],
        "properties": {
          "limit": {
            "type": "integer",
            "format": "int64"
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "plants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlantResponse"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  },
...
}